refactor ocr modules
